LayoutTests/fast/dom/Node/nullable-parameters-expected.txt

Tests the behavior of a few Node API parameters that should be nullable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.lookupPrefix('undefined') is "c"
PASS element.lookupPrefix(null) is null
PASS element.lookupPrefix(undefined) is null

PASS element.lookupNamespaceURI('undefined') is "test"
PASS element.lookupNamespaceURI(null) is "http://www.w3.org/1999/xhtml"
PASS element.lookupNamespaceURI(undefined) is "http://www.w3.org/1999/xhtml"

PASS element.isDefaultNamespace('undefined') is true
PASS element.isDefaultNamespace(undefined) is false
PASS element.isDefaultNamespace('null') is true
PASS element.isDefaultNamespace(null) is false
PASS successfullyParsed is true

TEST COMPLETE
